% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{convert_df_vcf}
\alias{convert_df_vcf}
\title{Convert a data.frame with VCF information to VCF text format}
\usage{
convert_df_vcf(df1, write_vcf = T, vcf_filename = NULL)
}
\arguments{
\item{df1}{data.frame with information about persons.
Only field names described in the \strong{Details} section  \code{\link{gen_vcf}} are converted}

\item{write_vcf}{Logical value indicating if information is written to \code{vcf_filename}}

\item{vcf_filename}{Character string with name of \code{.VCF} file to which the text information is written to
if \code{write_vcf==T}}
}
\value{
Character string with the information in VCF format if not written to file otherwise \code{NULL}
}
\description{
Convert a data.frame with VCF information to VCF text format
}
\examples{
\dontrun{
df1 <- tibble::tribble(
  ~fn , ~sn, ~phone_work ,
  "han", "oostdijk", "020-1234567",
  "john","smith","+31207654321")

convert_df_vcf(df1,write_vcf=F)
# BEGIN:VCARD
# VERSION:3.0
# N:oostdijk;han;;;
# FN:han oostdijk
# TEL;TYPE=WORK:020-1234567
# END:VCARD
# BEGIN:VCARD
# VERSION:3.0
# N:smith;john;;;
# FN:john smith
# TEL;TYPE=WORK:+31207654321
# END:VCARD

res <- convert_df_vcf(df1,write_vcf=T,vcf_filename="df1.vcf" )
is.null(res)
# [1] TRUE
}
}
