% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_registerpdf.R
\name{convert_reginfo_to_memadm}
\alias{convert_reginfo_to_memadm}
\title{convert registration info read from form to data.frame for member admin of TTV Amstelveen}
\usage{
convert_reginfo_to_memadm(df1)
}
\arguments{
\item{df1}{data.frame read by read_registerpdf}
}
\value{
A data.frame with the converted to format fit for member administration
}
\description{
convert registration info read from form to data.frame for member admin of TTV Amstelveen
}
\examples{
\dontrun{
df2 <- convert_reginfo_to_memadm(df1)
}
}
