% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{gen_vcf}
\alias{gen_vcf}
\title{Generate VCF information from list}
\usage{
gen_vcf(NL)
}
\arguments{
\item{NL}{Named list with information about a person. See Details}
}
\value{
character string with information for one person in VCF format
}
\description{
Generate VCF information from list
}
\section{Details}{

The named list can contain fields with the following names (case insensitive) while other fields will be ignored
\itemize{
\item fn : first name
\item sn : second name
\item group : character vector (or string with comma separated values) with group the person belongs to
\item email* : with * one of email, email_home, email_work or email_other
\item phone* : with * one of phone, phone_home, phone_home2, phone_work, phone_work2 or phone_mobile
\item street* : with * one of home work other
\item place* : with * one of home work other
\item state* : with * one of home work other
\item postcode* : with * one of home work other
\item country* : with * one of home work other
}
}

\examples{
\dontrun{
gen_vcf(
  list(fn="Han",sn="Oostdijk",
    street_work="Runstreet 1224",place_work="Amstelveen",
    postcode_work="1181XX",state_work="NH",country_work="NL",
    email_work="han@hanoostdijk.nl",phone_work="020-1234567")
 )
# BEGIN:VCARD
# VERSION:3.0
# N:Oostdijk;Han;;;
# FN:Han Oostdijk
# TEL;TYPE=WORK:020-1234567
# ADR;TYPE=WORK:;;Runstreet 1224;Amstelveen;NH;1181XX;NL
# END:VCARD
}
}
