% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_registerpdf.R
\name{read_registerpdf}
\alias{read_registerpdf}
\title{read pdf with registration info for TTV Amstelveen}
\usage{
read_registerpdf(filename)
}
\arguments{
\item{filename}{Character string with the name of the pdf file}
}
\value{
A data.frame with the fields of the pdf file: 1 row with 17 columns
}
\description{
read pdf with registration info for TTV Amstelveen
}
\examples{
\dontrun{
df1 <- read_registerpdf(filename)
}
}
