% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_registerpdf.R
\name{write_reginfo_to_memadm}
\alias{write_reginfo_to_memadm}
\title{write converted registration info to Excel sheet for member admin of TTV Amstelveen}
\usage{
write_reginfo_to_memadm(df1, wbname = "inschrijf.xlsx")
}
\arguments{
\item{df1}{data.frame created by convert_reginfo_to_memadm}

\item{wbname}{Character string with name of workbook to write to}
}
\value{
invisible(0)
}
\description{
write converted registration info to Excel sheet for member admin of TTV Amstelveen
}
\section{Details}{

The sheetname is "Nederlands" or "Engels" depending on the language of the (converted) registration form
}

\examples{
\dontrun{
df1 <- read_registerpdf(filename)
df2 <- convert_reginfo_to_memadm(df1)
write_reginfo_to_memadm(df2)
}
}
