% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_table.R
\name{create_html_table}
\alias{create_html_table}
\alias{create_style_table}
\title{create the html to display a data.frame}
\usage{
create_html_table(
  df1,
  header = list(names(df1)),
  tableclass = NULL,
  thclass = NULL,
  thstyle = NULL,
  tdclass = NULL,
  tdstyle = NULL,
  caption = NULL,
  captionclass = NULL,
  captionstyle = NULL,
  html_include = NULL
)

create_style_table(
  classname = NULL,
  tabborder = "2px solid black",
  border = "1px solid #ddd",
  width = "auto",
  padding_left = "4px",
  padding_top = padding_left,
  padding_right = padding_left,
  padding_bottom = padding_top,
  text_align = "center",
  background_color = "white",
  color = "black",
  white_space = NULL
)
}
\arguments{
\item{df1}{data.frame for which an html table representation will be made}

\item{header}{Character vector or list of character vectors or matrix of character string to use as th table headers. Each vector or matrix row generates one header row.}

\item{tableclass}{Character string with the class name to assign to the table tag}

\item{thclass}{Character string with the class name to assign to the th tags}

\item{thstyle}{Character string with the style information to assign to the th tags}

\item{tdclass}{Character string with the class name to assign to the td tags}

\item{tdstyle}{Character string with the style information to assign to the td tags}

\item{caption}{Character string with the table caption}

\item{captionclass}{Character string with the class for the table caption}

\item{captionstyle}{Character string with the style for the table caption}

\item{html_include}{Character string with html code that is included before the table statement}

\item{classname}{Character string with the class name to use in this style}

\item{tabborder}{Character string with \code{border} definition for \code{table}}

\item{border}{Character string with \code{border} definition for \code{th} and \code{td}}

\item{width}{Character string with \code{width} definition for \code{table}. E.g. \code{"auto"} or \code{"50\%"}}

\item{padding_left}{Character string with \code{padding-left} definition for \code{th} and \code{td}}

\item{padding_top}{Character string with \code{padding-top} definition for \code{th} and \code{td}}

\item{padding_right}{Character string with \code{padding-right} definition for \code{th} and \code{td}}

\item{padding_bottom}{Character string with \code{padding-bottom} definition for \code{th} and \code{td}}

\item{text_align}{Character string with \code{text-align} definition for \code{th} and \code{td}}

\item{background_color}{Character string with \code{background-color} definition for \code{th} and \code{td}}

\item{color}{Character string with \code{color} definition for \code{th} and \code{td}#'}

\item{white_space}{Character string with \code{white-space} definition for \code{th} and \code{td}}
}
\value{
For \code{create_html_table} a list() with a shiny.tag class that can be converted into an HTML string via \code{as.character()}, viewed with \code{htmltools::browsable()} and saved to a file with \code{htmltools::save_html()} .

For \code{create_style_table} a character string with the generated style block information.
}
\description{
The function \code{create_html_table} creates the html to display the rows of a data.frame in the table format. \cr
The header can be missing or contain one or more rows with the default being the names of the data.frame. \cr
A classname for the table and a caption with its class and style can be specified. \cr
Also for the \code{td} and \code{th} elements a class and style can be specified. However see the \code{create_style_table} function.\cr
It is possible to include a character string containing html information via the \code{html_include} argument.
This information can contain a \code{style} definition.\cr
\cr
The function \code{create_style_table} creates style information that can also be included by using the \code{html_include} argument of the \code{create_html_table} function.
This is an alternative for specifying the class and style for the \code{th} and \code{td} elements.
The function uses a style template that is filled in with the arguments of function. See \code{template_create_style_table} in the examples.
If the template does not satisfy your needs, you can specify your own template such as is done in the example \code{fit_table_def}.
}
\examples{
\dontrun{

# the style used in create_style_table is given below. The arguments of the function are glue-wise
# inserted between (^,$) pairs with exception of classname, width and white_space that get some preprocessing:
#  class1 <- ifelse(is.null(classname),"",paste0(".",classname))
#  width1 <- ifelse((is.null(width)||(nchar(width)==0)),"",paste0("width: ",width,";"))
#  ws1    <- ifelse((is.null(white_space)||(nchar(white_space)==0)),"",paste0("white-space: ",white_space,";"))
template_create_style_table <- "
<style>
  table^class1$ {
    border-collapse: collapse;
    border: ^tabborder$;
    table-layout: auto;
    ^width1$
  }

  ^class1$ :is(th, td) {
  border: ^border$;
  padding-top: ^padding_top$;
  padding-bottom: ^padding_bottom$;
  padding-left: ^padding_left$;
  padding-right: ^padding_right$;
  text-align: ^text_align$;
  background-color : ^background_color$ ;
  color :  ^color$ ;
  ^ws1$
}
</style>
  "

 fit_table_def <- "
 <style>
   table.fit-table {
    border: none;
    width: auto;
    table-layout: auto;
  }

   table.fit-table th, table.fit-table td {
    border: none;
    padding-top: 8px;
    padding-bottom: 8px;
    text-align: var(--tabel-align, left);
    white-space: nowrap;
  }

   .is-right {
     --tabel-align: right;
  }

   .is-center {
     --tabel-align: center;
  }
 </style>
  "

df1 <- data.frame(f1=c(1,2),f2=c("Antwerpen","Bern"))
tb1 <- create_html_table(df1)
tb1 <- create_html_table(df1,tableclass="fit-table is-center",html_include=fit_table_def,
    caption="this is my caption", captionstyle="white-space: nowrap;caption-side: bottom;")
htmltools::browsable(tb1)
cat( as.character(tb1) )
htmltools::save_html(tb1,'test.html')
}

\dontrun{
st <- create_style_table(classname="mc",tabborder="2px solid black", background_color = "#00ffFF",
   color="red",padding_top = "12px",padding_bottom = "2px")
tb <- create_html_table(df1,tableclass="mc",html_include=st)
}

}
