% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_table.R
\name{create_html_table}
\alias{create_html_table}
\title{create the html to display a data.frame}
\usage{
create_html_table(df1, header = list(names(df1)), class = NULL)
}
\arguments{
\item{df1}{data.frame for which an html table representation will be mad}

\item{header}{list of character vectors to use as th table headers. Each vector generate one header row.}

\item{class}{Character string with the class name to assign to the table statement}
}
\value{
A list() with a shiny.tag class that can be converted into an HTML string via as.character() and saved to a file with save_html().
}
\description{
create the html to display a data.frame
}
\examples{
\dontrun{
df1 <- data.frame(f1=c(1,2),f2=c("A","B"))
st <- create_html_table(df1)
}

}
