% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_html_table2.R
\name{create_html_table2}
\alias{create_html_table2}
\title{create the html to display a data.frame}
\usage{
create_html_table2(
  df1,
  headers = NULL,
  colspan = NULL,
  caption = NULL,
  tdHTML = F,
  style_info = NULL,
  class = NULL,
  html_file = NULL,
  format_thf = NULL,
  format_tdf = NULL
)
}
\arguments{
\item{df1}{data.frame for which an html table representation will be made}

\item{headers}{List of character vectors to use as th table headers.
Each vector generates one header row}

\item{colspan}{List of integer vectors of same length as \code{headers}
Each vector indicates for the corresponding header the column where a \code{colspan="2"} will be inserted in the \verb{th<>} definition}

\item{caption}{Character string to use as caption for the table}

\item{tdHTML}{Boolean indicating that (non-header) fields should be handled as HTML}

\item{style_info}{Character string with info that is included with the \verb{<style>} tag before the table statement.
Lines starting with the R comment symbol # will be removed}

\item{class}{Character string with the class name to assign to the table tag}

\item{html_file}{Character string which, if specified, indicates the filename in which the generated HTML is saved}

\item{format_thf}{Function with arguments \code{headers} and \code{colspan} that formats the th elements in a tr element}

\item{format_thd}{Function with arguments cell contents, \code{tdHTML} and optional column index that formats one td element}
}
\value{
A list() with a shiny.tag class that can be converted into an HTML string via \code{as.character()}, viewed with \code{htmltools::browsable()} and saved to a file with \code{htmltools::save_html()} .
}
\description{
The function \code{create_html_table2} creates the html to display the rows of a data.frame in HTML. \cr
The header can be missing or contain one or more rows with the default being the names of the data.frame. \cr
A caption can be specified and CSS information can be given in a character vector to be included with the \code{style_info} argument as a \verb{<style>} tag.\cr
}
\details{
\subsection{headers}{
\itemize{
\item the default headers are the column names of the data.frame: \code{names(df1)}
\item one or more header rows can be specified by the entries in a list: \code{list(c(...),c(...))}
\item no headers will be produced by specifying an empty list: \code{list()}
}
}
}
\examples{
create_html_table2(df1,html_file="h.html")


}
